backend/requirements.txt
PyPDF2==3.0.1
pytest==7.3.1
python-dotenv==1.0.0
pymongo>=4.5,<5
fastapi>=0.109.0,<0.110.0
uvicorn>=0.27.0,<0.28.0
black==25.1.0
flake8==7.2.0
pytest-asyncio==0.21.0
pycryptodome>=3.20.0
pdfplumber==0.11.6
pydantic>=2.0.0,<3.0.0
python-multipart==0.0.5
pytest-cov==4.1.0
httpx==0.24.1
redis==4.5.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor>=3.3.0
pytest-mock==3.10.0
aioredis==2.0.1 